+++ software/Seekfree_TC264_Opensource_Library/code/debug.h
@@ -9,8 +9,8 @@
 #define CODE_DEBUG_H_
 
 void LCDShowPerImage(void);//LCD显示透视后的图像
-void LCDShowFloatLine(myPoint_f* line,uint8 len,const uint16 color);//LCD显示边线
-void LCDShowUint8Line(myPoint* line,uint8 len,const uint16 color);//LCD显示边线
+void LCDShowFloatLine(myPoint_f* line,int len,const uint16 color);//LCD显示边线
+void LCDShowUint8Line(myPoint* line,int len,const uint16 color);//LCD显示边线
 
 
 #endif /* CODE_DEBUG_H_ */
